Flytt organisajon-service til GCP
