--- .github/workflows/deploy.organisasjon-api.dev.yml
@@ -1,48 +1,0 @@
-name: Deploy organisasjon-api-dev
-
-on:
-  push:
-    paths:
-      - apps/organisasjon-api/**
-      - libs/testnorge-core/**
-
-env:
-  IMAGE: docker.pkg.github.com/${{ github.repository }}/organisasjon-api:${{ github.sha }}
-
-jobs:
-  build_and_deploy:
-    name: Build and deploy
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v1
-      - name: Gradle cache
-        uses: actions/cache@v2
-        with:
-          path: |
-            ~/.gradle
-          key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle*') }}
-          restore-keys: |
-            ${{ runner.os }}-gradle-
-      - uses: actions/setup-java@v1
-        with:
-          java-version: '11.x'
-      - name: Compile
-        env:
-          GITHUB_USERNAME: token
-          NAV_TOKEN: ${{ secrets.NAV_TOKEN }}
-        run: ./gradlew :organisasjon-api:build
-      - name: Build and publish Docker image
-        working-directory: apps/organisasjon-api/
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          docker build --tag $IMAGE .
-          docker login docker.pkg.github.com -u $GITHUB_REPOSITORY -p $GITHUB_TOKEN
-          docker push $IMAGE
-      - name: Deploy to nais
-        timeout-minutes: 5
-        uses: nais/deploy/actions/deploy@v1
-        env:
-          APIKEY: ${{ secrets.NAIS_DOLLY_DEPLOY_API_KEY }}
-          CLUSTER: dev-fss
-          RESOURCE: apps/organisasjon-api/config.dev.yml
