--- apps/organisasjon-api/config.yml
@@ -1,51 +1,0 @@
-﻿apiVersion: "nais.io/v1alpha1"
-kind: "Application"
-metadata:
-  name: testnorge-organisasjon-api
-  namespace: dolly
-  labels:
-    team: dolly
-spec:
-  image: {{ image }}
-  port: 8080
-  webproxy: true
-  azure:
-    application:
-      enabled: true
-      tenant: nav.no
-  accessPolicy:
-    inbound:
-      rules:
-        - application: team-dolly-lokal-app
-          cluster: dev-fss
-        - application: mn-organisasjon-api
-        - application: testnorge-statisk-data-forvalter
-        - application: organisasjon-orgnummer-service
-        - application: testnorge-synt-sykemelding-api
-        - application: organisasjon-forvalter
-        - application: testnorge-oversikt-frontend
-  liveness:
-    path: /internal/isAlive
-    initialDelay: 4
-    periodSeconds: 5
-    failureThreshold: 500
-  readiness:
-    path: /internal/isReady
-    initialDelay: 4
-    periodSeconds: 5
-    failureThreshold: 500
-  replicas:
-    min: 1
-    max: 1
-  vault:
-    enabled: true
-  resources:
-    requests:
-      cpu: 200m
-      memory: 4000Mi
-    limits:
-      cpu: 500m
-      memory: 5000Mi
-  ingresses:
-    - "https://testnorge-organisasjon-api.dev.adeo.no"
-    - "https://testnorge-organisasjon-api.dev.intern.nav.no"
