--- apps/organisasjon-api/src/main/java/no/nav/registre/testnorge/organisasjon/OrganisasjonApiApplicationStarter.java
@@ -1,11 +1,0 @@
-package no.nav.registre.testnorge.organisasjon;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrganisasjonApiApplicationStarter {
-    public static void main(String[] args) {
-        SpringApplication.run(OrganisasjonApiApplicationStarter.class, args);
-    }
-}
