--- apps/organisasjon-api/src/main/java/no/nav/registre/testnorge/organisasjon/config/ApplicationConfig.java
@@ -1,13 +1,0 @@
-package no.nav.registre.testnorge.organisasjon.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-import no.nav.registre.testnorge.libs.core.config.ApplicationCoreConfig;
-
-@Configuration
-@Import({
-        ApplicationCoreConfig.class
-})
-public class ApplicationConfig {
-}
