--- apps/organisasjon-api/src/main/java/no/nav/registre/testnorge/organisasjon/config/DevConfig.java
@@ -1,13 +1,0 @@
-package no.nav.registre.testnorge.organisasjon.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Profile;
-
-import no.nav.registre.testnorge.libs.localdevelopment.LocalDevelopmentConfig;
-
-@Configuration
-@Profile("dev")
-@Import(LocalDevelopmentConfig.class)
-public class DevConfig {
-}
