--- apps/organisasjon-api/src/main/java/no/nav/registre/testnorge/organisasjon/consumer/dto/DetaljerDTO.java
@@ -1,19 +1,0 @@
-package no.nav.registre.testnorge.organisasjon.consumer.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-@Value
-@NoArgsConstructor(force = true)
-@AllArgsConstructor
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class DetaljerDTO {
-    @JsonProperty
-    private String enhetstype;
-}
-
