--- apps/organisasjon-api/src/main/java/no/nav/registre/testnorge/organisasjon/consumer/dto/NavnDTO.java
@@ -1,21 +1,0 @@
-package no.nav.registre.testnorge.organisasjon.consumer.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-@Value
-@Builder
-@NoArgsConstructor(force = true)
-@AllArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class NavnDTO {
-
-    @JsonProperty
-    private String redigertnavn;
-    @JsonProperty
-    private String navnelinje1;
-}
