--- apps/organisasjon-api/src/main/java/no/nav/registre/testnorge/organisasjon/consumer/dto/OrganisasjonDetaljerDTO.java
@@ -1,22 +1,0 @@
-package no.nav.registre.testnorge.organisasjon.consumer.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-import java.util.List;
-
-@Value
-@NoArgsConstructor(force = true)
-@AllArgsConstructor
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class OrganisasjonDetaljerDTO {
-    @JsonProperty
-    private final List<AdresseDTO> forretningsadresser;
-    @JsonProperty
-    private final List<AdresseDTO> postadresser;
-}
