--- apps/organisasjon-api/src/main/java/no/nav/registre/testnorge/organisasjon/service/OrganisasjonService.java
@@ -1,17 +1,0 @@
-package no.nav.registre.testnorge.organisasjon.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import no.nav.registre.testnorge.organisasjon.consumer.EregConsumer;
-import no.nav.registre.testnorge.organisasjon.domain.Organisasjon;
-
-@Service
-@RequiredArgsConstructor
-public class OrganisasjonService {
-    private final EregConsumer consumer;
-
-    public Organisasjon getOrganisasjon(String orgnummer, String miljo) {
-        return consumer.getOrganisasjon(orgnummer, miljo);
-    }
-}
