--- apps/organisasjon-api/src/main/resources/application.yml
@@ -1,13 +1,0 @@
-spring:
-  application:
-    name: testnorge-organisasjon-api
-    version: application.version.todo #TODO Finn ut hvordan denne kan settes fra gradle
-    description: "Testnorge organisasjoner API"
-
-
-AAD_ISSUER_URI: https://login.microsoftonline.com/62366534-1ec3-4962-8869-9b5535279d0b
-
-
-consumers:
-  ereg:
-    url: https://modapp-{miljo}.adeo.no/ereg/api/v1/
