--- apps/organisasjon-api/src/test/java/no/nav/registre/testnorge/organisasjon/ApplicationContextTest.java
@@ -1,25 +1,0 @@
-package no.nav.registre.testnorge.organisasjon;
-
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit4.SpringRunner;
-
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureWireMock(port = 0)
-@TestPropertySource(locations = "classpath:application-test.properties")
-public class ApplicationContextTest {
-
-    /**
-     * Used to test app runtime dependency and properties
-     */
-    @Test
-    @SuppressWarnings("java:S2699")
-    public void load_app_context() {
-    }
-}
