--- apps/organisasjon-api/src/test/resources/application-test.properties
@@ -1,3 +1,0 @@
-consumers.ereg.url=http://localhost:${wiremock.server.port}/ereg-{miljo}/api/v1/
-
-spring.cloud.vault.token=dummy
