+++ apps/organisasjon-service/build.gradle
@@ -1,0 +1,35 @@
+plugins {
+    id 'no.nav.registre.testnorge.java-conventions'
+    id 'org.springframework.boot' version '2.3.4.RELEASE'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+}
+bootJar {
+    archiveFileName = "app.jar"
+}
+
+dependencies {
+    implementation project(':local-development')
+    implementation project(':testnorge-core')
+    implementation project(':testnorge-oauth2')
+    implementation project(':testnorge-data-transfer-objects')
+    implementation project(':testnorge-common-command')
+
+    implementation 'org.springframework.security:spring-security-web:5.3.4.RELEASE'
+    implementation 'org.springframework.security:spring-security-config:5.3.4.RELEASE'
+    implementation 'org.springframework.security:spring-security-core:5.3.4.RELEASE'
+    implementation 'org.springframework.boot:spring-boot-starter-oauth2-client:2.3.4.RELEASE'
+    implementation 'org.springframework.boot:spring-boot-starter-webflux:2.3.2.RELEASE'
+    implementation 'org.hibernate.validator:hibernate-validator:6.1.5.Final'
+    implementation 'org.springframework.cloud:spring-cloud-starter-vault-config:2.2.5.RELEASE'
+    implementation 'org.springframework.cloud:spring-cloud-starter-sleuth:2.2.5.RELEASE'
+    implementation 'org.springframework.boot:spring-boot-starter-web:2.3.4.RELEASE'
+    implementation 'org.springdoc:springdoc-openapi-ui:1.4.6'
+    implementation 'net.logstash.logback:logstash-logback-encoder:6.3'
+    implementation 'org.projectlombok:lombok:1.18.12'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.3.4.RELEASE'
+    testImplementation 'org.springframework.cloud:spring-cloud-contract-wiremock:2.2.4.RELEASE'
+    testImplementation project(':testnorge-testing')
+    testImplementation 'org.assertj:assertj-core:3.16.1'
+}
+
+description = 'organisasjon-service'
