+++ apps/organisasjon-service/src/main/java/no/nav/registre/testnorge/organisasjonservice/OrganisasjonServiceApplicationStarter.java
@@ -1,0 +1,11 @@
+package no.nav.registre.testnorge.organisasjonservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OrganisasjonServiceApplicationStarter {
+    public static void main(String[] args) {
+        SpringApplication.run(OrganisasjonServiceApplicationStarter.class, args);
+    }
+}
