+++ apps/organisasjon-service/src/main/java/no/nav/registre/testnorge/organisasjonservice/config/ApplicationConfig.java
@@ -1,0 +1,15 @@
+package no.nav.registre.testnorge.organisasjonservice.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+import no.nav.registre.testnorge.libs.core.config.ApplicationCoreConfig;
+import no.nav.registre.testnorge.libs.oauth2.config.SecureOAuth2ServerToServerConfiguration;
+
+@Configuration
+@Import({
+        ApplicationCoreConfig.class,
+        SecureOAuth2ServerToServerConfiguration.class
+})
+public class ApplicationConfig {
+}
