+++ apps/organisasjon-service/src/main/java/no/nav/registre/testnorge/organisasjonservice/config/DevConfig.java
@@ -1,0 +1,13 @@
+package no.nav.registre.testnorge.organisasjonservice.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Profile;
+
+import no.nav.registre.testnorge.libs.localdevelopment.LocalDevelopmentConfig;
+
+@Configuration
+@Profile("dev")
+@Import(LocalDevelopmentConfig.class)
+public class DevConfig {
+}
