+++ apps/organisasjon-service/src/main/java/no/nav/registre/testnorge/organisasjonservice/consumer/dto/NavnDTO.java
@@ -1,0 +1,21 @@
+package no.nav.registre.testnorge.organisasjonservice.consumer.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+import lombok.Value;
+
+@Value
+@Builder
+@NoArgsConstructor(force = true)
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class NavnDTO {
+
+    @JsonProperty
+    String redigertnavn;
+    @JsonProperty
+    String navnelinje1;
+}
