+++ apps/organisasjon-service/src/main/java/no/nav/registre/testnorge/organisasjonservice/provider/OrganisasjonController.java
@@ -1,0 +1,34 @@
+package no.nav.registre.testnorge.organisasjonservice.provider;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import no.nav.registre.testnorge.libs.dto.organisasjon.v1.OrganisasjonDTO;
+import no.nav.registre.testnorge.organisasjonservice.service.OrganisasjonService;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/organisasjoner")
+public class OrganisasjonController {
+    private final OrganisasjonService organisasjonService;
+
+    @GetMapping("/{orgnummer}")
+    public ResponseEntity<OrganisasjonDTO> getOrgnaisasjon(
+            @PathVariable("orgnummer") String orgnummer,
+            @RequestHeader("miljo") String miljo
+    ) {
+        log.info("Henter organisasjon {} fra {}.", orgnummer, miljo);
+        var organisasjon = organisasjonService.getOrganisasjon(orgnummer, miljo);
+        if (organisasjon == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(organisasjon.toDTO());
+    }
+}
