+++ apps/organisasjon-service/src/main/resources/bootstrap-prod.yml
