+++ apps/organisasjon-service/src/test/resources/application-test.properties
@@ -1,0 +1,3 @@
+consumers.ereg.url=http://localhost:${wiremock.server.port}
+
+spring.cloud.vault.token=dummy
