+++ settings.gradle
@@ -76,7 +76,7 @@ include(':testnorge-avro-schema')
 include(':mn-organisasjon-api')
 include(':organisasjon-api')
 include(':testnorge-database')
-include(':rapportering-api')
+include(':organisasjon-service')
 include(':token-provider-core')
 include(':arbeidsforhold-xsd')
 include(':helsepersonell-service')
@@ -167,7 +167,7 @@ project(":testnorge-avro-schema").projectDir = file('libs/testnorge-avro-schema'
 project(":mn-organisasjon-api").projectDir = file('apps/mn-organisasjon-api')
 project(":organisasjon-api").projectDir = file('apps/organisasjon-api')
 project(":testnorge-database").projectDir = file('libs/testnorge-database')
-project(":rapportering-api").projectDir = file('apps/rapportering-api')
+project(":organisasjon-service").projectDir = file('apps/organisasjon-service')
 project(":token-provider-core").projectDir = file('apps/testnorge-token-provider/token-provider-core')
 project(":arbeidsforhold-xsd").projectDir = file('xsd/arbeidsforhold-xsd')
 project(":helsepersonell-service").projectDir = file('apps/helsepersonell-service')
